Fix TCK ControllerConstraintHandlerTest
When we run the ControllerConstraintHandlerTest the body is read twice.
Once from the BodyBinder via the request inputstream and then again from the On Error handler.

The issue is that without changes, the second read gets nothing as the inputstream is exhausted.
And we cannot cache the stream internally, as it is expected that the body is of the same type
as created by the binder.

This fix introduces an interface so that the binder can effectively cache the body in the request.
This cached body is then returned on the subsequent calls to getBody()