Stop double encoding body when we have a response encoder
When a response-encoder is present, it handles encoding the response.

Without this change, we were then falling through and encoding the body into the response a second time.

This was discovered in AWS here https://github.com/micronaut-projects/micronaut-aws/issues/1843

It wasn't seen here as we add a servlet for handling static resources in jetty/undertow/tomcat