Add Servlet Context if deployed as named WAR file
If we are deploying as a named WAR file (ie: `myproject-0.1.war`) then this will be the context path for the application.
This change detects this, and sets the context path for the application accordingly.
Any context-path config in the application context will be ignored in this case.

If the deployed WAR file is added to the ROOT context (ie: by deploying `ROOT.war`), then any context-path from application properties will be used.