GCN-4404: Initial commit of serverless micronaut web applications.
This is currently using rawhttp as parser. I will change it soon to use a more widely used parser.

I am able to test a sample Micronaut application with this module. The sample can be found in samples/ dir along with its own README and build script.
I can see the app working both in interactive mode as well as inetd mode.

Here's the plan before PR is submitted:
1. Use a more widely used parser.
2. Add test cases and samples for http-serverless module.
3. Fix TODOs in code.
4. Run compatibility tests for micronaut.
5. Update gcn/micronaut launcher to allow users to select the new serverless option.